Force a rebuild on aarch64.
If we don’t, it seems that a build of the kernel will include x86_64
binaries into the image :S. Ideally going forward we’ll detect which
image was last built, and use that to determine if a rebuild is
necessary. For now play it safe.

This doesn’t appear to be an issue for x86_64.