Swap qemu escape key from C-a to C-b.
In the terminal C-a is a shortcut for 'go to start of line', which makes this
choice of shortcut extremely unfortate and an error I've constantly made while
using qemu in -nographic mode. C-b has no such issues.

I had to ask in the irc channel to find this out :) seems poorly documented out
there on the internet.